requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
# Authentication
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ML libraries
scikit-learn
pandas
numpy
joblib

# Validation
pydantic
pydantic-settings
email-validator

# Testing (optional)
pytest
pytest-asyncio
httpx

python-dotenv
psycopg2-binary